Split nested types into separate files.
